Middleware를 이용한 인증 처리
- '/' 경로에 접근할 때 middleware에서 로그인 상태에 따라 /login 또는 /dashboard로 리디렉션을 처리
- 폴더 구조 이동
