+++ src/app/(auth)/components/LoginPrompt.tsx
