+++ src/app/(auth)/components/SignupPrompt.tsx
